Fix Var Name Mismatch And Ensure We Don't Attempt To Access Nil Vals
This update adds a little bit of protection to the status display
page -- ensuring that if the status.name var isn't set or the
status.message var isn't set that we don't try to access their members.
It also ensures that the correct var names are used in each place
when attempting to find the length of the string.