Add support for specifying UUID
If you pass :uuid => "foo" into a job's `.create` call, resque-status will use
that UUID instead of generating a new one.

If a job with that UUID already exists, we check to see if it is in a
"killable" state, meaning it is yet-to-happen or happening right now. If so,
we refuse to queue and return nil.  If it is completed, crashed, etc, we delete
the old job and queue the new one.